intelligraph-bot/backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
python-dotenv==1.0.0
requests==2.31.0
openai==1.3.7
llama-index==0.9.13
langchain==0.0.350
neo4j==5.15.0
spacy==3.7.2
pandas==2.1.4
numpy==1.25.2
beautifulsoup4==4.12.2
scikit-learn==1.3.2
transformers==4.36.2
torch==2.1.1
networkx==3.2.1
rdflib==7.0.0
supabase==2.3.0
arxiv==2.0.0
scholarly==1.7.11
python-multipart==0.0.6
uvicorn==0.24.0
fastapi==0.104.1
pydantic==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
